feat: Add Cornish-Fisher VaR risk management, FinBERT sentiment pipeline, and fix pre-existing resilience hierarchy halt bugs
- Add risk.rs module with Cornish-Fisher VaR computation (99% confidence)
- Add sentiment.rs module with FinBERT sentiment extraction (BGE-small-en-v1.5)
- Extend ArbitrationInput with sentiment_score and sentiment_confidence fields
- Upgrade tri_level_validator.rs to 4-layer parallel validation (Rules 35%, ML 25%, Chronos 25%, Sentiment 15%)
- Add VaR emergency gate that forces HOLD when risk tolerance exceeded
- Fix hardcoded database paths to use StorageConfig for cross-platform compatibility
- Fix resilience hierarchy test: use manual_halt for SymbolSpecific severity
- Fix outcome_processor test: use unique in-memory DB to avoid lock contention
- Add architecture.md and workflow.md professional documentation
- Remove auto-generated audit files
- Update .gitignore for comprehensive coverage

All 323 tests passing (181 autonomous + 129 core + 13 ML)

+++ crates/cotrader-core/src/patterns.rs
@@ -1220,7 +1220,7 @@ fn detect_pin_bar(bars: &[OhlcvBar]) -> Option<CandlestickPattern> {
         return Some(CandlestickPattern {
             name: "Bullish Pin Bar".to_string(),
             direction: "bullish".to_string(),
-            strength: strength.clamp(0.0, 0.9),
+            strength: (strength as f64).clamp(0.0, 0.9),
             bar_index: bars.len() - 1,
         });
     }
@@ -1231,7 +1231,7 @@ fn detect_pin_bar(bars: &[OhlcvBar]) -> Option<CandlestickPattern> {
         return Some(CandlestickPattern {
             name: "Bearish Pin Bar".to_string(),
             direction: "bearish".to_string(),
-            strength: strength.clamp(0.0, 0.9),
+            strength: (strength as f64).clamp(0.0, 0.9),
             bar_index: bars.len() - 1,
         });
     }
@@ -1259,7 +1259,7 @@ fn detect_fakey(bars: &[OhlcvBar]) -> Option<CandlestickPattern> {
         return Some(CandlestickPattern {
             name: "Bullish Fakey".to_string(),
             direction: "bullish".to_string(),
-            strength: strength.clamp(0.0, 0.9),
+            strength: (strength as f64).clamp(0.0, 0.9),
             bar_index: bars.len() - 1,
         });
     }
@@ -1270,7 +1270,7 @@ fn detect_fakey(bars: &[OhlcvBar]) -> Option<CandlestickPattern> {
         return Some(CandlestickPattern {
             name: "Bearish Fakey".to_string(),
             direction: "bearish".to_string(),
-            strength: strength.clamp(0.0, 0.9),
+            strength: (strength as f64).clamp(0.0, 0.9),
             bar_index: bars.len() - 1,
         });
     }
